COMMERCE-4870 commerce-machine-learning-impl Move Batch resource mapper to own package
