+++ modules/dxp/apps/commerce/commerce-machine-learning-impl/src/main/java/com/liferay/commerce/machine/learning/internal/forecast/dispatch/AnalyticsDownloadCommerceAccountCommerceMLForecastDispatchTaskExecutor.java
@@ -14,7 +14,7 @@
 
 package com.liferay.commerce.machine.learning.internal.forecast.dispatch;
 
-import com.liferay.commerce.machine.learning.internal.data.integration.BatchEngineTaskItemDelegateResourceMapper;
+import com.liferay.commerce.machine.learning.internal.batch.BatchEngineTaskItemDelegateResourceMapper;
 import com.liferay.commerce.machine.learning.internal.dispatch.AnalyticsDispatchTaskExecutor;
 import com.liferay.dispatch.executor.BaseDispatchTaskExecutor;
 import com.liferay.dispatch.executor.DispatchTaskExecutor;
